src/requirements.txt
WTForms~=2.3.3
Pillow~=7.2.0
Flask~=1.1.2
itsdangerous~=1.1.0
python-dotenv~=0.15.0
email_validator
flask-bcrypt
flask-login